+++ src/ClientInterface.php
@@ -5,10 +5,13 @@ namespace alexeevdv\SumSub;
 use alexeevdv\SumSub\Exception\Exception;
 use alexeevdv\SumSub\Request\AccessTokenRequest;
 use alexeevdv\SumSub\Request\ApplicantDataRequest;
+use alexeevdv\SumSub\Request\ApplicantStatusRequest;
+use alexeevdv\SumSub\Request\DocumentImagesRequest;
 use alexeevdv\SumSub\Request\ResetApplicantRequest;
 use alexeevdv\SumSub\Response\AccessTokenResponse;
 use alexeevdv\SumSub\Response\ApplicantDataResponse;
-use alexeevdv\SumSub\Response\ResetApplicantResponse;
+use alexeevdv\SumSub\Response\ApplicantStatusResponse;
+use alexeevdv\SumSub\Response\DocumentImagesResponse;
 
 interface ClientInterface
 {
@@ -32,4 +35,18 @@ interface ClientInterface
      * @throws Exception
      */
     public function resetApplicant(ResetApplicantRequest $request): void;
+
+    /**
+     * Get applicant status
+     * @see https://developers.sumsub.com/api-reference/#getting-applicant-status-api
+     * @throws Exception
+     */
+    public function getApplicantStatus(ApplicantStatusRequest $request): ApplicantStatusResponse;
+
+    /**
+     * Get document images
+     * @see https://developers.sumsub.com/api-reference/#getting-document-images
+     * @throws Exception
+     */
+    public function getDocumentImages(DocumentImagesRequest $request): DocumentImagesResponse;
 }