+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Under development
 -----------------
 
+0.4.0 (2022-04-21)
+-----------------
+- Feature: getApplicantStatus() method support (sspat)
+- Feature: getDocumentImages() method support (sspat)
+
 0.3.0 (2022-04-20)
 -----------------
 - Feature: resetApplicant() method support (sspat)